Calling Obj-C method from Kotlin-JVM
Now Kotlin runtime should ship with a .dylib with a bunch of native methods for
calling Objective-C from java. All native methods right now are placed in
jet.runtime.objc.Native class, and the ordinary methods called from the mirror
classes are put in jet.runtime.objc.ObjC class. In order to build a .dylib
for now, run runtime/native/make_native script (this should be placed in the
project buildfile later)

Support the simplest case: a void method in class object with no arguments

Due to the lack of any Java or Kotlin libraries, the only viable way to test
that Objective-C is actually called right now is through native printf and
comparing the text output with the hard-coded expected output