Support Obj-C methods returning unknown classes
Objective-C method which is declared to return NSString *, can instead return
an object of some class which is unexposed at compile time, e.g.
__NSCFConstantString. When constructing a Java object from such a return value,
we now go up the hierarchy to find the first class which is available at
compile time (there should be at least one) and create an instance of it.