Retain Obj-C objects
If an object is returned by some Objective-C method, we need to tell
Objective-C Runtime that we now have an additional reference to it (in the
mirroring object in JVM). To prevent memory leaks, we should also release the
object somewhere, e.g. generate the corresponding code in the 'finalize()'
method of every mirroring JVM class