Obj-C metaclasses inheritance
Class objects of Obj-C classes inherit from the root class of the class of
which they're in. Because of this, a fake root class 'NSObject' was added to
the most of the resolve tests not to duplicate dat (otherwise class members
of the root class would appear twice, as declarations in the class and as fake
overrides in its class object).

There's no such thing as a class object of a protocol or a category, so don't
resolve them anymore (but still resolve metaclasses for protocols/categories,
since the latter may contain class methods)