Support built-in Obj-C types in return value position
id is now resolved as ObjCObject, SEL as ObjCSelector, Class as ObjCClass.

Native.objc_msgSend now is the only message-sending native method, and it
returns a NativeValue, instance of which is created in native code based on the
signature of the called method. Returned primitives are stored into
PrimitiveValue, which now has a few useful accessor methods.

Problem with BOOL resolving as id is fixed by including objc/objc.h, where it's
declared.

Also some actions were taken for tests to reduce used memory in tests